cronjob for inactive user and api enhancements

+++ comment/views.py
@@ -55,7 +55,7 @@ class FollowUnfollowView(generics.GenericAPIView):
       
 class ListAllFollowersView(generics.ListAPIView):
     serializer_class = FollowersDetailSerializer
-    # permission_class = [permissions.IsAuthenticated]
+    permission_class = [permissions.IsAuthenticated]
     
     def get_queryset(self):
         id = self.kwargs.get('id')
@@ -63,7 +63,7 @@ class ListAllFollowersView(generics.ListAPIView):
 
 class ListAllFollowingView(generics.ListAPIView):
     serializer_class = FollowersDetailSerializer
-    # permission_class = [permissions.IsAuthenticated]
+    permission_class = [permissions.IsAuthenticated]
     
     def get_queryset(self):
         id = self.kwargs.get('id')
@@ -78,8 +78,10 @@ class AlbumLikeDislikeView(generics.CreateAPIView, generics.DestroyAPIView):
     
     def delete(self, request):
       try :
-        AlbumLikes.objects.get(album_id = request.data['album_id'], user_id = request.data['user_id']).delete()
-        return Response({'message' : 'request successful'}, status = 200)
+        if AlbumLikes.objects.filter(album_id = request.data['album_id'], user_id = request.data['user_id']).exists():
+          AlbumLikes.objects.get(album_id = request.data['album_id'], user_id = request.data['user_id']).delete()
+          return Response({'message' : 'request successful'}, status = 200)
+        return Response({'message' : 'already disliked'}, status = 201)
       
       except Exception as e:
         return Response({'message' : str(e)}, status = 400)
@@ -93,9 +95,10 @@ class PlaylistLikeDislikeView(generics.CreateAPIView, generics.DestroyAPIView):
     
     def delete(self, request):
       try :
-        PlaylistLikes.objects.get(playlist_id = request.data['playlist_id'], user_id = request.data['user_id']).delete()
-        return Response({'message' : 'request successful'}, status = 200)
-      
+        if PlaylistLikes.objects.filter(playlist_id = request.data['playlist_id'], user_id = request.data['user_id']).exists() :
+          PlaylistLikes.objects.get(playlist_id = request.data['playlist_id'], user_id = request.data['user_id']).delete()
+          return Response({'message' : 'request successful'}, status = 200)
+        return Response({'message' : 'already disliked'}, status = 201)
       except Exception as e:
         return Response({'message' : str(e)}, status = 400)
       
@@ -108,8 +111,9 @@ class SongLikeDislikeView(generics.CreateAPIView, generics.DestroyAPIView):
     
     def delete(self, request):
       try :
-        SongLikes.objects.get(song_id = request.data['song_id'], user_id = request.data['user_id']).delete()
-        return Response({'message' : 'request successful'}, status = 200)
-      
+        if SongLikes.objects.filter(song_id = request.data['song_id'], user_id = request.data['user_id']).exists():
+          SongLikes.objects.get(song_id = request.data['song_id'], user_id = request.data['user_id']).delete()
+          return Response({'message' : 'request successful'}, status = 200)
+        return Response({'message' : 'already disliked'}, status = 201)
       except Exception as e:
         return Response({'message' : str(e)}, status = 400)
